Introduce interface for calculators
